fix(Alex review): split `serverpod_test` files better, remove tx timeout
