+++ packages/serverpod_test/lib/serverpod_test.dart
@@ -1,3 +1,8 @@
 library serverpod_test;
 
-export 'src/serverpod_test.dart';
+export 'src/with_serverpod.dart';
+export 'src/test_session.dart';
+
+// Only used by generated code
+export 'src/test_serverpod.dart';
+export 'src/function_call_wrappers.dart';
