+++ packages/serverpod_test/lib/src/test_serverpod.dart
@@ -1,0 +1,49 @@
+import 'package:serverpod/serverpod.dart';
+
+/// Internal test endpoints interface that contains implementation details
+/// that should only be used internally by the test tools.
+abstract interface class InternalTestEndpoints {
+  /// Initializes the endpoints.
+  void initialize(
+    SerializationManagerServer serializationManager,
+    EndpointDispatch endpoints,
+  );
+}
+
+/// A facade for the real Serverpod instance.
+class TestServerpod<T extends InternalTestEndpoints> {
+  /// The test endpoints that are exposed to the user.
+  T testEndpoints;
+
+  final Serverpod _serverpod;
+
+  /// Creates a new test serverpod instance.
+  TestServerpod({
+    required this.testEndpoints,
+    required SerializationManagerServer serializationManager,
+    required EndpointDispatch endpoints,
+    String? runMode,
+  }) : _serverpod = Serverpod(
+          ['-m', runMode ?? ServerpodRunMode.test],
+          serializationManager,
+          endpoints,
+        ) {
+    endpoints.initializeEndpoints(_serverpod.server);
+    testEndpoints.initialize(serializationManager, endpoints);
+  }
+
+  /// Starts the underlying serverpod instance.
+  Future<void> start() async {
+    await _serverpod.start();
+  }
+
+  /// Shuts down the underlying serverpod instance.
+  Future<void> shutdown() async {
+    return _serverpod.shutdown(exitProcess: false);
+  }
+
+  /// Creates a new Serverpod session.
+  Future<Session> createSession({bool enableLogging = false}) async {
+    return _serverpod.createSession(enableLogging: enableLogging);
+  }
+}
